third_party/blink/web_tests/platform/generic/http/tests/security/contentTypeOptions/nosniff-script-without-content-type-blocked-expected.txt

CONSOLE ERROR: Refused to execute script from 'http://127.0.0.1:8000/security/contentTypeOptions/resources/script-with-header.pl' because its MIME type ('') is not executable, and strict MIME type checking is enabled.
Check that script sent with an 'X-Content-Type-Options: nosniff' header is correctly blocked if no 'Content-Type' header is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.scriptsSuccessfullyLoaded is 0
PASS successfullyParsed is true

TEST COMPLETE
